+++ inst/NEWS.md
@@ -9,6 +9,7 @@
 
 * Remove debug javascript messages
 * Fix javascript error for lenses which activated wrong columns
+* Fix bug when no n_cells or n_features is entered
 
 # dynguidelines 0.3.0 (15-11-2018)